ТЗ 8 на вторую проверку

+++ src/main/java/ru/yandex/group/filmorate/model/User.java
@@ -12,8 +12,7 @@ import java.time.LocalDate;
 @Data
 @AllArgsConstructor
 public class User {
-    private int id;
-
+    private long id;
     @Email
     private String email;
